# public/swagger_v4.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "url": "https://api.bible.build/v2",
      "description": "Online Staging server for testing"
    },
    {
      "url": "https://api.dbp.dev/v2",
      "description": "Internal Staging server for testing"
    }
  ],
  "info": {
    "title": "Digital Bible Platform [Koinos]",
    "description": "This is a simple API",
    "version": "4.0.0",
    "contact": {
      "name": "API Support",
      "url": "https://dbs.org/contact",
      "email": "jon@dbs.org"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "paths": {
...
    "/bibles/filesets/{id}/permissions": {
      "get": {
        "operationId": "v4_bible_filesets.permissions_index",
        "summary": "Get a list of permissions for all filesets.",
        "parameters": [
          {
            "$ref": "#/components/parameters/v"
          },
          {
            "$ref": "#/components/parameters/key"
          },
          {
            "$ref": "#/components/parameters/version"
          },
          {
            "$ref": "#/components/parameters/reply"
          },
          {
            "$ref": "#/components/parameters/pretty"
          }
        ],
        "tags": [
          "Bible"
        ],
        "responses": {
          "200": {
            "description": "A set of Access details for a specific fileset",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "data": {
                      "hash_id": "006af44d95cc",
                      "fileset_id": "HUUTBLN1DA",
                      "whitelist": true,
                      "access_granted": true,
                      "granted_at": "2018-04-17 06:06:18",
                      "updated_at": "2018-04-17 06:06:18"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "v4_bible_filesets.permissions_store",
        "summary": "To add a new set of permissions to the API.",
        "parameters": [
          {
            "$ref": "#/components/parameters/v"
          },
          {
            "$ref": "#/components/parameters/key"
          },
          {
            "$ref": "#/components/parameters/version"
          },
          {
            "$ref": "#/components/parameters/reply"
          },
          {
            "$ref": "#/components/parameters/pretty"
          }
        ],
        "tags": [
          "Bible"
        ],
        "responses": {
          "200": {
            "description": "You must either be a super_admin or be the owner of the fileset to be authorized for this route.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "No fileset has been found for the provided hash_id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "error": {
                      "message": "No fileset has been found for the provided hash_id",
                      "status_code": 404
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "v4_bible_filesets.permissions_update",
        "summary": "To update an existing set of permissions to the API.",
        "parameters": [
          {
            "$ref": "#/components/parameters/v"
          },
          {
            "$ref": "#/components/parameters/key"
          },
          {
            "$ref": "#/components/parameters/version"
          },
          {
            "$ref": "#/components/parameters/reply"
          },
          {
            "$ref": "#/components/parameters/pretty"
          }
        ],
        "tags": [
          "Bible"
        ],
        "responses": {
          "200": {
            "description": "You must either be a admin or be the owner of the fileset to be authorized for this route.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": null
                }
              }
            }
          }
        }
      }
    },
...
    "/resources/{resource_id}": {
      "get": {
        "operationId": "v4_resources.show",
        "summary": "Get a single resource by it's ID",
        "parameters": [
          {
            "$ref": "#/components/parameters/v"
          },
          {
            "$ref": "#/components/parameters/key"
          },
          {
            "$ref": "#/components/parameters/version"
          },
          {
            "$ref": "#/components/parameters/reply"
          },
          {
            "$ref": "#/components/parameters/pretty"
          }
        ],
        "tags": [
          "Organizations"
        ],
        "responses": {
          "200": {
            "description": "A success resource index query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resource"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "v4_resources.update",
        "summary": "Update a single resource by it's ID.",
        "description": "This route will update a single resource by it's ID but in order to do so you must either belong to the organization that owns the resource or have `archivist` level permissions. Some resources are updated via API. Changes made to this type of resource will be overwritten on the next update. These resources are indentified by a `source_id`",
        "parameters": [
          {
            "$ref": "#/components/parameters/v"
          },
          {
            "$ref": "#/components/parameters/key"
          },
          {
            "$ref": "#/components/parameters/version"
          },
          {
            "$ref": "#/components/parameters/reply"
          },
          {
            "$ref": "#/components/parameters/pretty"
          }
        ],
        "tags": [
          "Organizations"
        ],
        "responses": {
          "200": {
            "description": "A successful resource update query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resource"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "v4_resources.store",
        "summary": "Create a single resource",
        "description": "This route will create a single resource but order to do so you must either belong to an organization or have `archivist` level permissions. Please do not create a resource for which your organization already has a connected API as that will introduce duplicates into the system. To see a list of connected APIs, please see the `connections` route.",
        "parameters": [
          {
            "$ref": "#/components/parameters/v"
          },
          {
            "$ref": "#/components/parameters/key"
          },
          {
            "$ref": "#/components/parameters/version"
          },
          {
            "$ref": "#/components/parameters/reply"
          },
          {
            "$ref": "#/components/parameters/pretty"
          }
        ],
        "tags": [
          "Organizations"
        ],
        "responses": {
          "200": {
            "description": "A successful resource creation query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resource"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "v4_resources.destroy",
        "summary": "Delete a Single Resource",
        "description": "This route will delete a single resource but order to do so you must either belong to the owning organization or have `archivist` level permissions.",
        "parameters": [
          {
            "$ref": "#/components/parameters/v"
          },
          {
            "$ref": "#/components/parameters/key"
          },
          {
            "$ref": "#/components/parameters/version"
          },
          {
            "$ref": "#/components/parameters/reply"
          },
          {
            "$ref": "#/components/parameters/pretty"
          }
        ],
        "tags": [
          "Organizations"
        ],
        "responses": {
          "200": {
            "description": "A successful deletion query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resource"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Resource": {
        "required": [
          "id",
          "name",
          "glotto_id",
          "iso"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "uniqueItems": true,
            "description": "An incrementing ID for the resources",
            "example": 1
          },
          "iso": {
            "type": "string",
            "description": "An iso 639-3 code to identify the primary language of the resource",
            "example": "eng"
          },
          "cover": {
            "type": "string",
            "example": "https://example.com/image.png"
          },
          "cover_thumnail": {
            "type": "string",
            "example": "https://example.com/image_thumbnail.png"
          },
          "translations": {
            "type": "object",
            "example": {
              "eng": {
                "title": "The Story of the Home for Bible Translators",
                "description": "The Story of the HfBT"
              },
              "spa": {
                "title": "बाइबल अनुवादकहरूको लागि घर",
                "description": "The Story of the HfBT"
              }
            }
          },
          "organization": {
            "type": "object",
            "example": {
              "slug": "the-home-for-bible-translators",
              "primaryColor": "#293e80",
              "secondaryColor": "#FFFFFF",
              "translations": {
                "nep": "बाइबल अनुवादकहरूको लागि घर",
                "spa": "El Hogar para Traductores Bíblicos",
                "eng": "The Home for Bible Translators"
              }
            }
          },
          "derived": {
            "type": "string",
            "example": "ENGKJV"
          },
          "copyright": {
            "type": "string",
            "example": "Public Domain"
          },
          "links": {
            "type": "object",
            "example": [
              {
                "title": "Store",
                "size": "246 pages",
                "type": "Print",
                "url": "http://www.librarything.com/work/1301652/book/118270355"
              },
              {
                "title": "PDF Download",
                "size": "25.4 mb",
                "url": "https://example.com/download.pdf"
              }
            ]
          }
        }
      },
...
    },
    "parameters": {
      "v": {
        "name": "v",
        "in": "query",
        "description": "The Version Number of the API",
        "required": true,
        "example": 4,
        "schema": {
          "type": "string"
        }
      },
      "key": {
        "name": "key",
        "in": "query",
        "description": "The API key that's been generated for you.",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "pretty": {
        "name": "pretty",
        "in": "query",
        "description": "A sugar param, simply will format the call for human eyes",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "version": {
        "name": "key",
        "in": "query",
        "description": "The API key that's been generated for you.",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "reply": {
        "name": "reply",
        "in": "query",
        "description": "Specifies the response type requested by the caller. Default is json.",
        "schema": {
          "type": "string",
          "enum": [
            "json",
            "html",
            "jsonp",
            "yaml",
            "xml"
          ]
        }
      },
...
    }
  }
}